Rename the instrumentation entry point
`instrumentation-client.ts` is a filename Next recognised and loaded on
its own. Nothing loads it by name any more, `main.tsx` imports it, so
the name was pointing at a mechanism that no longer exists.
